+++ src/main/java/janggi/domain/moveRules/MaMoveRule.java
@@ -1,0 +1,23 @@
+package janggi.domain.moveRules;
+
+import janggi.domain.Direction;
+import janggi.domain.MoveRule;
+import janggi.domain.Route;
+import janggi.domain.Team;
+import java.util.List;
+
+public class MaMoveRule implements MoveRule {
+
+    @Override
+    public List<Route> findRoutes(Team team) {
+            Route route1 = new Route(List.of(Direction.UP, Direction.UP_LEFT));
+            Route route2 = new Route( List.of(Direction.UP, Direction.UP_RIGHT));
+            Route route3 = new Route(List.of(Direction.RIGHT, Direction.UP_RIGHT));
+            Route route4 = new Route(List.of(Direction.RIGHT, Direction.DOWN_RIGHT));
+            Route route5 = new Route(List.of(Direction.DOWN, Direction.DOWN_RIGHT));
+            Route route6 = new Route( List.of(Direction.DOWN, Direction.DOWN_LEFT));
+            Route route7 = new Route( List.of(Direction.LEFT, Direction.DOWN_LEFT));
+            Route route8 = new Route( List.of(Direction.LEFT, Direction.UP_LEFT));
+            return List.of(route1, route2, route3, route4, route5, route6, route7, route8);
+    }
+}
